requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
pydantic==2.10.0
requests==2.32.3
chardet==6.0.0
numpy==2.1.0
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-docx==1.1.2
odfpy==1.4.1
sse-starlette==2.1.3
pywebview==5.1
psutil==6.1.1
GPUtil==1.4.0
rich==13.9.4
pyinstaller==6.11.0
